+++ crates/resolver/src/lookup_ip.rs
@@ -27,8 +27,7 @@ use crate::error::*;
 use crate::hosts::Hosts;
 use crate::lookup::{Lookup, LookupEither, LookupIntoIter, LookupIter};
 use crate::lookup_state::CachingClient;
-use crate::name_server::{Connection, StandardConnection};
-use crate::TokioSpawnBg;
+use crate::name_server::{TokioConnection, TokioConnectionProvider};
 
 /// Result of a DNS query when querying for A or AAAA records.
 ///
@@ -107,7 +106,7 @@ impl Iterator for LookupIpIntoIter {
 /// The Future returned from [`AsyncResolver`] when performing an A or AAAA lookup.
 ///
 /// This type isn't necessarily something that should be used by users, see the default TypeParameters are generally correct
-pub struct LookupIpFuture<C = LookupEither<Connection, StandardConnection, TokioSpawnBg>>
+pub struct LookupIpFuture<C = LookupEither<TokioConnection, TokioConnectionProvider>>
 where
     C: DnsHandle + 'static,
 {